# perf: optimize test execution speed with fast test configuration