## Summary
- Add fast test configuration for 65% faster local development and pre-commit hooks
- Increase thread pool utilization from 2-4 to 10 threads for 12-core machine  
- Maintain full coverage reporting for CI/CD pipelines

## Performance Improvements
- **Full test suite**: 40s → 14s (⚡ **65% faster**)
- **Unit tests**: 34s → 10s (⚡ **69% faster**)
- **Pre-commit hooks**: Now 65% faster for better developer experience

## Changes Made
- ✅ Created `vitest.config.fast.ts` without coverage collection
- ✅ Added fast test scripts: `test:fast`, `test:unit:fast`, `test:integration:fast`
- ✅ Updated pre-commit hooks to use fast tests
- ✅ Created fast integration test runner for Electron environment
- ✅ Updated CLAUDE.md with new command documentation

## New Commands Available
```bash
# Fast tests for local development (no coverage)
npm run test:fast              # ~14s vs ~40s
npm run test:unit:fast         # ~10s vs ~34s  
npm run test:integration:fast  # ~2s

# Coverage tests for CI/CD (unchanged)
npm run test                   # Full coverage reporting
npm run test:unit             # Unit tests with coverage
npm run test:integration      # Integration tests with coverage
```

## Test plan
- [x] Verify fast tests run successfully without coverage
- [x] Verify original test commands still work with full coverage
- [x] Verify pre-commit hooks use fast tests and complete faster
- [x] Confirm all 2,187 unit tests + 115 integration tests pass
- [x] Ensure TypeScript compilation and linting pass
- [x] Verify build process completes successfully